Move Sep31Info and Sep38Info to org.stellar.anchor.api.asset package
