+++ api-schema/src/main/java/org/stellar/anchor/api/asset/Sep31Info.java
@@ -1,9 +1,8 @@
-package org.stellar.anchor.api.sep.operation;
+package org.stellar.anchor.api.asset;
 
 import com.google.gson.annotations.SerializedName;
 import java.util.Map;
 import lombok.Data;
-import org.stellar.anchor.api.asset.AssetInfo;
 
 @Data
 public class Sep31Info {
