+++ api-schema/src/main/java/org/stellar/anchor/api/asset/StellarAssetInfo.java
@@ -2,8 +2,6 @@ package org.stellar.anchor.api.asset;
 
 import com.google.gson.annotations.SerializedName;
 import lombok.Data;
-import org.stellar.anchor.api.sep.operation.Sep31Info;
-import org.stellar.anchor.api.sep.operation.Sep38Info;
 
 @Data
 public class StellarAssetInfo implements AssetInfo {
