+++ api-schema/src/main/java/org/stellar/anchor/api/sep/sep31/Sep31InfoResponse.java
@@ -3,7 +3,7 @@ package org.stellar.anchor.api.sep.sep31;
 import com.google.gson.annotations.SerializedName;
 import java.util.Map;
 import lombok.Data;
-import org.stellar.anchor.api.sep.operation.Sep31Info;
+import org.stellar.anchor.api.asset.Sep31Info;
 
 /**
  * The response body of the /info endpoint of the SEP-31.
